Changed to version 1.1.3 Buffed osmium chances Added osmium and bronze bow

+++ src/main/java/net/kuliiuu/metalsplus/util/ModModelPredicates.java
@@ -17,7 +17,7 @@ public class ModModelPredicates {
             if (entity == null) {
                 return 0.0F;
             } else {
-                return entity.getActiveItem() != stack ? 0.0F : (stack.getMaxUseTime(entity) - entity.getItemUseTimeLeft()) / 20.0F;
+                return entity.getActiveItem() != stack ? 0.0F : (float)(stack.getMaxUseTime(entity) - entity.getItemUseTimeLeft()) / 20.0F;
             }
         });
         ModelPredicateProviderRegistry.register(
